% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_book.R
\name{pdf_book_fnaufel}
\alias{pdf_book_fnaufel}
\title{pdf_book_fnaufel}
\usage{
pdf_book_fnaufel(
  toc = TRUE,
  number_sections = TRUE,
  fig_caption = TRUE,
  pandoc_args = paste("--variable=link-as-notes", paste0("--variable=", "resources:",
    system.file("rmarkdown/resources", package = "bd.format")),
    "--variable=subparagraph", paste0("--filter=",
    system.file("rmarkdown/resources/filters/dollar.py3", package = "bd.format")),
    paste0("--filter=", system.file("rmarkdown/resources/filters/hl.py3", package =
    "bd.format"))),
  base_format = rmarkdown::pdf_document,
  toc_unnumbered = TRUE,
  toc_appendix = TRUE,
  toc_bib = TRUE,
  quote_footer = NULL,
  highlight_bw = FALSE,
  df_print = "tibble",
  keep_tex = TRUE,
  latex_engine = "xelatex",
  citation_package = "natbib",
  includes = rmarkdown::includes(in_header =
    system.file("rmarkdown/resources/00-preambulo.tex", package = "bd.format")),
  md_extensions = "+fenced_divs+bracketed_spans",
  template = system.file("rmarkdown/resources/bd.format.latex", package = "bd.format"),
  ...
)
}
\arguments{
\item{toc}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{number_sections}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{fig_caption}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{pandoc_args}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{base_format}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{toc_unnumbered}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{toc_appendix}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{toc_bib}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{quote_footer}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{highlight_bw}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{df_print}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{keep_tex}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{latex_engine}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{citation_package}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{includes}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{md_extensions}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{template}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}

\item{...}{see \link[bookdown:pdf_book]{bookdown::pdf_book}}
}
\value{
see \link[bookdown:pdf_book]{bookdown::pdf_book}
}
\description{
Custom PDF template
}
\details{
Loads additional style and configs
}
\author{
fnaufel
}
